src/main/java/com/jam2in/arcus/admin/tool/bean/UserDetailsService.java
package com.jam2in.arcus.admin.tool.bean;

import com.jam2in.arcus.admin.tool.domain.user.entity.UserEntity;
import com.jam2in.arcus.admin.tool.domain.user.repository.UserRepository;
import org.apache.commons.collections4.CollectionUtils;
import org.apache.commons.lang3.StringUtils;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.transaction.annotation.Transactional;

import java.util.stream.Collectors;
import java.util.stream.Stream;

public class UserDetailsService
    implements org.springframework.security.core.userdetails.UserDetailsService {

  private final UserRepository userRepository;

  public UserDetailsService(UserRepository userRepository) {
    this.userRepository = userRepository;
  }

  @Override
  @Transactional(readOnly = true)
  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
    UserEntity userEntity = userRepository.findByUsername(username)
        .orElseThrow(() -> new UsernameNotFoundException(StringUtils.EMPTY));

    return new User(
        userEntity.getUsername(),
        userEntity.getPassword(),
        Stream.concat(
            Stream.of(userEntity.getRole()),
            CollectionUtils.emptyIfNull(userEntity.getAccesses()).stream())
            .collect(Collectors.toList()));
  }

}
